requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cpu
matplotlib==3.10.7
numpy==2.3.3
scikit-learn==1.7.2
tqdm==4.67.1
torch==2.9.0
# torch==2.8.0+cpu
torchvision==0.24.0
# torchvision==0.23.0+cpu
subspyces==0.0.5
ipywidgets==8.1.7
cocopp==2.8.0
coco-experiment==2.8.1
requests==2.32.5